fix: ensure school_id_govt is well formatted and reuse code
